requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-groq>=0.1.0
langchain-huggingface>=0.0.3
sentence-transformers
python-dotenv
fastapi
uvicorn
